fix tiny compiler warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@2011 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtlib/core/CustomTerrain.cpp
@@ -19,7 +19,7 @@
 // (You don't have to use them, especially for height)
 //
 #define MAKE_XYZ1(x,y) m_fXLookup[(x)], m_pData[offset((x),(y))]*m_fZScale, m_fZLookup[(y)]
-#define MAKE_XYZ2(x,y) (x), m_pData[offset((x),(y))], (y)
+#define MAKE_XYZ2(x,y) (float)(x), m_pData[offset((x),(y))], (float)(y)
 
 
 //
